setsketcher: make slog dependency optional
The upstream argmin crate uses the slog feature dynamic keys which makes
it a little tricky to incorporate in projects that already use slog with
default features (because it changes the type signatures of most
things).

This patch proposes making the slog integration optional here.

Tested as follows:

```
cargo check
cargo check --no-default-features
```